refactor: decompose execute_operation into focused modules
Extract AST, markdown, tidy, and patch operation handlers from the
monolithic execute_operation() match arms in tx/execute.rs into
dedicated submodules:

- tx/ast_op.rs (~400 lines): AST list, read, rename, validate ops
- tx/md_op.rs (~170 lines): markdown section, bullet, table, dedupe,
  lint, insert ops
- tx/tidy_op.rs (~55 lines): tidy normalize op
- tx/patch_op.rs (~45 lines): patch apply op

execute_operation() now delegates to these modules via thin match arms,
reducing its size and improving navigability.

Closes #1359
