+++ src/tx/execute.rs
@@ -1,16 +1,10 @@
 use super::output::{TxExecResult, TxLintResult, TxReadResult, TxSearchResult};
 use super::validate::op_label;
-use crate::cli::global::{EolMode, GlobalFlags};
+use crate::cli::global::GlobalFlags;
 use crate::ops::doc::{
     FileFormat, MutationResult, apply_doc_mutation, detect_format, parse_doc,
     serialize_value_preserving,
 };
-use crate::ops::md::{
-    dedupe_headings_in, insert_after_heading_in, insert_before_heading_in, move_section_in,
-    replace_section_in, upsert_bullet_in,
-};
-use crate::ops::patch::{ApplyHunksOptions, ApplyHunksStatus, apply_patch_with_loader};
-
 use crate::plan::{Operation, Plan};
 use crate::write::{WritePolicy, apply_policy};
 
@@ -101,34 +95,6 @@ pub(crate) fn mark_write_target(write_targets: &mut HashSet<PathBuf>, path: &Pat
 }
 
 // ---------------------------------------------------------------------------
-// AST directory expansion (#1287)
-// ---------------------------------------------------------------------------
-
-/// Walk a directory and collect files that have a tree-sitter grammar.
-/// Used by ast.rename/ast.replace in execute_plan when the path is a directory.
-#[cfg(feature = "ast")]
-fn collect_ast_source_files(dir: &Path) -> anyhow::Result<Vec<PathBuf>> {
-    let mut files = Vec::new();
-    collect_ast_source_files_recursive(dir, &mut files)?;
-    files.sort();
-    Ok(files)
-}
-
-#[cfg(feature = "ast")]
-fn collect_ast_source_files_recursive(dir: &Path, out: &mut Vec<PathBuf>) -> anyhow::Result<()> {
-    for entry in std::fs::read_dir(dir)? {
-        let entry = entry?;
-        let path = entry.path();
-        if path.is_dir() {
-            collect_ast_source_files_recursive(&path, out)?;
-        } else if path.is_file() && crate::ast::Language::from_path(&path).has_grammar() {
-            out.push(path);
-        }
-    }
-    Ok(())
-}
-
-// ---------------------------------------------------------------------------
 // Operation execution
 // ---------------------------------------------------------------------------
 
@@ -162,35 +128,6 @@ pub(crate) fn flush_doc_cache(
     for (path, cached) in doc_cache.drain() {
         flush_doc_cache_entry(pending, deletions, write_targets, path, cached)?;
     }
-    Ok(())
-}
-
-/// Apply a markdown heading operation (read file, transform, write back).
-///
-/// Five of the six md operations follow an identical pattern: resolve path,
-/// read content, call a `(&str, &str, &str) -> Option<String>` transform,
-/// error on `None`, and update pending state. This helper captures that pattern.
-pub(crate) fn apply_md_heading_op(
-    tx: &mut TxState<'_>,
-    path: &str,
-    heading: &str,
-    extra: &str,
-    op: impl FnOnce(&str, &str, &str) -> Option<String>,
-    err_label: &str,
-) -> anyhow::Result<()> {
-    let file_path = tx.cwd.join(path);
-    let file_content = read_file_content(tx.pending, tx.existed_before, &file_path)?;
-    let new_content =
-        op(file_content, heading, extra).ok_or_else(|| crate::exit::NoMatchError {
-            msg: format!("{err_label} not found: {heading}"),
-        })?;
-    update_file_content(
-        tx.pending,
-        tx.deletions,
-        tx.write_targets,
-        &file_path,
-        new_content,
-    );
     Ok(())
 }
 
@@ -621,57 +558,6 @@ pub(crate) fn execute_file_op(op: &Operation, tx: &mut TxState<'_>) -> anyhow::R
     Ok(0)
 }
 
-/// Rename identifiers in a single file using AST-aware renaming with
-/// word-boundary fallback. Used by the AstRename handler (both single-file
-/// and directory expansion paths).
-#[cfg(feature = "ast")]
-fn ast_rename_single_file(
-    tx: &mut TxState<'_>,
-    abs: &Path,
-    old_name: &str,
-    new_name: &str,
-    lang_hint: Option<&str>,
-) -> anyhow::Result<usize> {
-    let content = read_file_content(tx.pending, tx.existed_before, abs)?;
-    let lang_val = lang_hint
-        .map(crate::ast::Language::from_name_or_ext)
-        .unwrap_or_else(|| crate::ast::Language::from_path(abs));
-    let result = crate::ast::rename::rename_in_source(content, old_name, new_name, lang_val);
-    match result {
-        Some(r) if r.replacements > 0 => {
-            update_file_content(tx.pending, tx.deletions, tx.write_targets, abs, r.content);
-            Ok(r.replacements)
-        }
-        Some(_) => Err(crate::exit::NoMatchError {
-            msg: format!("no matches for '{}' in {}", old_name, abs.display()),
-        }
-        .into()),
-        None => {
-            // Tree-sitter couldn't parse. Fallback to word-boundary replace.
-            let re =
-                crate::ops::replace::compile_replace_regex(old_name, false, false, false, true)?;
-            if let Some(re) = re {
-                let new_content = re.replace_all(content, new_name).to_string();
-                let count = re.find_iter(content).count();
-                if count > 0 {
-                    update_file_content(
-                        tx.pending,
-                        tx.deletions,
-                        tx.write_targets,
-                        abs,
-                        new_content,
-                    );
-                    return Ok(count);
-                }
-            }
-            Err(crate::exit::NoMatchError {
-                msg: format!("no matches for '{}' in {}", old_name, abs.display()),
-            }
-            .into())
-        }
-    }
-}
-
 pub(crate) fn execute_operation(op: &Operation, tx: &mut TxState<'_>) -> anyhow::Result<usize> {
     // Guard is enforced upfront in execute_plan_direct (for library plans) and via MCP pre-checks.
     // Single-op api::* uses ensure_contained inside write paths.
@@ -693,175 +579,21 @@ pub(crate) fn execute_operation(op: &Operation, tx: &mut TxState<'_>) -> anyhow:
             execute_doc_op(op, tx)?;
         }
 
-        Operation::MdReplaceSection {
-            path,
-            heading,
-            content,
-        } => {
-            apply_md_heading_op(tx, path, heading, content, replace_section_in, "heading")?;
+        // execute_md_op extracted to tx/md_op.rs (#1359).
+        Operation::MdReplaceSection { .. }
+        | Operation::MdInsertAfterHeading { .. }
+        | Operation::MdInsertBeforeHeading { .. }
+        | Operation::MdUpsertBullet { .. }
+        | Operation::MdTableAppend { .. }
+        | Operation::MdMoveSection { .. }
+        | Operation::MdDedupeHeadings { .. }
+        | Operation::MdLintAgents { .. } => {
+            return super::md_op::execute_md_op(op, tx);
         }
 
-        Operation::MdInsertAfterHeading {
-            path,
-            heading,
-            content,
-        } => {
-            apply_md_heading_op(
-                tx,
-                path,
-                heading,
-                content,
-                insert_after_heading_in,
-                "heading",
-            )?;
-        }
-
-        Operation::MdInsertBeforeHeading {
-            path,
-            heading,
-            content,
-        } => {
-            apply_md_heading_op(
-                tx,
-                path,
-                heading,
-                content,
-                insert_before_heading_in,
-                "heading",
-            )?;
-        }
-
-        Operation::MdUpsertBullet {
-            path,
-            heading,
-            bullet,
-        } => {
-            apply_md_heading_op(tx, path, heading, bullet, upsert_bullet_in, "heading")?;
-        }
-
-        Operation::MdTableAppend { path, heading, row } => {
-            let file_path = tx.cwd.join(path);
-            let file_content = read_file_content(tx.pending, tx.existed_before, &file_path)?;
-            let (body_start, body_end) = crate::ops::md::find_section(file_content, heading)
-                .ok_or_else(|| crate::exit::NoMatchError {
-                    msg: format!("heading not found: {heading}"),
-                })?;
-            let new_content =
-                crate::ops::md::table_append_in(file_content, body_start, body_end, row)
-                    .map_err(|e| anyhow::anyhow!("{e} under heading {heading:?}"))?;
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &file_path,
-                new_content,
-            );
-        }
-
-        Operation::MdMoveSection {
-            path,
-            heading,
-            to,
-            before,
-            after,
-        } => {
-            let position = match (before.as_deref(), after.as_deref()) {
-                (Some(b), None) => ("before", b),
-                (None, Some(a)) => ("after", a),
-                _ => anyhow::bail!("md.move_section requires exactly one of 'before' or 'after'"),
-            };
-            let dest_path_str = to.as_deref().unwrap_or(path.as_str());
-            let source_path = tx.cwd.join(path);
-            let dest_path = tx.cwd.join(dest_path_str);
-            let same_file = to.is_none()
-                || source_path == dest_path
-                || matches!(
-                    (source_path.canonicalize(), dest_path.canonicalize()),
-                    (Ok(ref s), Ok(ref d)) if s == d
-                );
-            let source_content =
-                read_file_content(tx.pending, tx.existed_before, &source_path)?.to_owned();
-            let dest_content = if same_file {
-                source_content.clone()
-            } else {
-                read_file_content(tx.pending, tx.existed_before, &dest_path)?.to_owned()
-            };
-            let (new_source, new_dest) =
-                move_section_in(&source_content, heading, &dest_content, position, same_file)
-                    .ok_or_else(|| crate::exit::NoMatchError {
-                        msg: "md.move_section: heading or target not found".to_string(),
-                    })?;
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &source_path,
-                new_source,
-            );
-            if !same_file {
-                update_file_content(
-                    tx.pending,
-                    tx.deletions,
-                    tx.write_targets,
-                    &dest_path,
-                    new_dest,
-                );
-            }
-        }
-
-        Operation::MdDedupeHeadings { path } => {
-            let file_path = tx.cwd.join(path);
-            let file_content = read_file_content(tx.pending, tx.existed_before, &file_path)?;
-            let (new_content, _removed) = dedupe_headings_in(file_content);
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &file_path,
-                new_content,
-            );
-        }
-
-        Operation::TidyFix {
-            path,
-            ensure_final_newline,
-            trim_trailing_whitespace,
-            normalize_eol,
-            collapse_blanks,
-            dedent,
-            indent,
-            lines,
-        } => {
-            let file_path = tx.cwd.join(path);
-            mark_write_target(tx.write_targets, &file_path);
-            let content = read_file_content(tx.pending, tx.existed_before, &file_path)?.to_owned();
-            let policy = WritePolicy {
-                ensure_final_newline: ensure_final_newline.unwrap_or(true),
-                trim_trailing_whitespace: trim_trailing_whitespace.unwrap_or(false),
-                normalize_eol: if let Some(eol) = normalize_eol {
-                    crate::write::parse_eol_mode(eol)?
-                } else {
-                    EolMode::Keep
-                },
-                collapse_blanks: collapse_blanks.unwrap_or(false),
-            };
-            let mut new = crate::write::apply_policy(&content, &policy).into_owned();
-
-            // Apply dedent/indent after policy normalization.
-            let line_range = lines
-                .as_deref()
-                .map(crate::ops::read::parse_line_range)
-                .transpose()?;
-            if let Some(spec) = dedent {
-                new = crate::write::dedent_content(&new, spec, line_range);
-            }
-            if let Some(spec) = indent {
-                new = crate::write::indent_content(&new, spec, line_range);
-            }
-
-            if content != new {
-                update_file_content(tx.pending, tx.deletions, tx.write_targets, &file_path, new);
-            }
+        // execute_tidy_op extracted to tx/tidy_op.rs (#1359).
+        Operation::TidyFix { .. } => {
+            return super::tidy_op::execute_tidy_op(op, tx);
         }
 
         Operation::FileAppend { .. }
@@ -872,46 +604,9 @@ pub(crate) fn execute_operation(op: &Operation, tx: &mut TxState<'_>) -> anyhow:
             return execute_file_op(op, tx);
         }
 
-        Operation::PatchApply {
-            diff,
-            on_stale,
-            allow_conflicts,
-        } => {
-            let options = ApplyHunksOptions {
-                on_stale: *on_stale,
-                allow_conflicts: *allow_conflicts,
-            };
-            let patched_files = apply_patch_with_loader(
-                diff,
-                |path| {
-                    let file_path = tx.cwd.join(path);
-                    Ok(read_file_content(tx.pending, tx.existed_before, &file_path)?.to_string())
-                },
-                options,
-            )?;
-            for result in patched_files {
-                if result.status == ApplyHunksStatus::Conflict && !allow_conflicts {
-                    anyhow::bail!(
-                        "patch apply: {} -- merge produced {} conflict(s); set allow_conflicts to write conflict markers",
-                        result.path,
-                        result.conflicts.len()
-                    );
-                }
-                let file_path = tx.cwd.join(&result.path);
-                if result.is_deletion {
-                    // File deletion via patch: mark for deletion.
-                    tx.deletions.insert(file_path.clone());
-                    tx.write_targets.insert(file_path);
-                } else {
-                    update_file_content(
-                        tx.pending,
-                        tx.deletions,
-                        tx.write_targets,
-                        &file_path,
-                        result.content,
-                    );
-                }
-            }
+        // execute_patch_op extracted to tx/patch_op.rs (#1359).
+        Operation::PatchApply { .. } => {
+            return super::patch_op::execute_patch_op(op, tx);
         }
 
         Operation::Read { path, lines } => {
@@ -922,441 +617,18 @@ pub(crate) fn execute_operation(op: &Operation, tx: &mut TxState<'_>) -> anyhow:
             execute_search_op(op, tx)?;
         }
 
-        Operation::MdLintAgents { path } => {
-            let file_path = tx.cwd.join(path);
-            let content = read_file_content(tx.pending, tx.existed_before, &file_path)?;
-            let issues = crate::ops::md::lint_agents_content(content);
-            tx.tx_lints.push(TxLintResult {
-                path: path.clone(),
-                issue_count: issues.len(),
-                issues,
-            });
-        }
-
         #[cfg(feature = "ast")]
-        Operation::AstRename {
-            path,
-            old_name,
-            new_name,
-            lang,
-        } => {
-            let abs = tx.cwd.join(path);
-
-            // Directory support (#1287): expand to individual files.
-            if abs.is_dir() {
-                let files = collect_ast_source_files(&abs)?;
-                if files.is_empty() {
-                    anyhow::bail!("no source files found in {}", path);
-                }
-                let mut total = 0usize;
-                for file_path in &files {
-                    let count =
-                        ast_rename_single_file(tx, file_path, old_name, new_name, lang.as_deref());
-                    if let Ok(n) = count {
-                        total += n;
-                    }
-                }
-                if total == 0 {
-                    return Err(crate::exit::NoMatchError {
-                        msg: format!("no matches for '{}' in {}", old_name, path),
-                    }
-                    .into());
-                }
-                return Ok(total);
-            }
-
-            return ast_rename_single_file(tx, &abs, old_name, new_name, lang.as_deref());
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstReplace {
-            path,
-            symbol,
-            old,
-            new_text,
-            regex,
-            lang,
-        } => {
-            let abs = tx.cwd.join(path);
-            let content = read_file_content(tx.pending, tx.existed_before, &abs)?;
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs));
-            let result = crate::ast::replace::replace_in_symbol(
-                content, symbol, old, new_text, *regex, lang_val,
-            )?;
-            match result {
-                Some(r) if r.replacements > 0 => {
-                    update_file_content(
-                        tx.pending,
-                        tx.deletions,
-                        tx.write_targets,
-                        &abs,
-                        r.content,
-                    );
-                    return Ok(r.replacements);
-                }
-                Some(_) => {
-                    return Err(crate::exit::NoMatchError {
-                        msg: format!(
-                            "no matches for '{}' in symbol '{}' in {}",
-                            old, symbol, path
-                        ),
-                    }
-                    .into());
-                }
-                None => {
-                    return Err(crate::exit::NoMatchError {
-                        msg: format!("symbol '{}' not found in {}", symbol, path),
-                    }
-                    .into());
-                }
-            }
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstInsert {
-            path,
-            content: insert_content,
-            inside,
-            after,
-            before,
-            position,
-            lang,
-        } => {
-            let abs = tx.cwd.join(path);
-            let file_content = read_file_content(tx.pending, tx.existed_before, &abs)?;
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs));
-            let pos = match position.as_deref() {
-                Some("start") => crate::ast::insert::InsertPosition::Start,
-                _ => crate::ast::insert::InsertPosition::End,
-            };
-            let result = crate::ast::insert::insert_code(
-                file_content,
-                insert_content,
-                inside.as_deref(),
-                after.as_deref(),
-                before.as_deref(),
-                pos,
-                lang_val,
-            )?;
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &abs,
-                result.content,
-            );
-            return Ok(1);
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstWrap {
-            path,
-            symbols: sym_names,
-            lines: line_range,
-            wrapper,
-            preamble,
-            lang,
-        } => {
-            let abs = tx.cwd.join(path);
-            let file_content = read_file_content(tx.pending, tx.existed_before, &abs)?;
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs));
-            let result = crate::ast::wrap::wrap_code(
-                file_content,
-                sym_names.as_deref(),
-                line_range.as_deref(),
-                wrapper,
-                preamble.as_deref(),
-                lang_val,
-            )?;
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &abs,
-                result.content,
-            );
-            return Ok(1);
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstImports {
-            path,
-            add,
-            remove,
-            dedupe,
-            lang,
-        } => {
-            let abs = tx.cwd.join(path);
-            let mut file_content =
-                read_file_content(tx.pending, tx.existed_before, &abs)?.to_string();
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs));
-            let mut total_changes = 0usize;
-
-            if let Some(add_list) = add {
-                let result = crate::ast::imports::add_imports(&file_content, add_list, lang_val);
-                total_changes += result.added;
-                file_content = result.content;
-            }
-            if let Some(remove_list) = remove {
-                let result =
-                    crate::ast::imports::remove_imports(&file_content, remove_list, lang_val);
-                total_changes += result.removed;
-                file_content = result.content;
-            }
-            if *dedupe {
-                let result = crate::ast::imports::dedupe_imports(&file_content, lang_val);
-                total_changes += result.deduped;
-                file_content = result.content;
-            }
-            if total_changes > 0 {
-                update_file_content(
-                    tx.pending,
-                    tx.deletions,
-                    tx.write_targets,
-                    &abs,
-                    file_content,
-                );
-            }
-            return Ok(total_changes);
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstReorder {
-            path,
-            inside,
-            order,
-            lang,
-        } => {
-            let abs = tx.cwd.join(path);
-            let file_content = read_file_content(tx.pending, tx.existed_before, &abs)?;
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs));
-            let strategy = crate::ast::reorder::parse_strategy(order)?;
-            let result = crate::ast::reorder::reorder_symbols(
-                file_content,
-                inside.as_deref(),
-                &strategy,
-                lang_val,
-            )?;
-            if result.symbols_reordered > 0 {
-                update_file_content(
-                    tx.pending,
-                    tx.deletions,
-                    tx.write_targets,
-                    &abs,
-                    result.content,
-                );
-            }
-            return Ok(result.symbols_reordered);
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstGroup {
-            path,
-            module,
-            symbols: sym_names,
-            preamble,
-            position,
-            lang,
-        } => {
-            let abs = tx.cwd.join(path);
-            let file_content = read_file_content(tx.pending, tx.existed_before, &abs)?;
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs));
-            let pos = match position.as_deref() {
-                Some("end") => crate::ast::group::GroupPosition::End,
-                Some(s) if let Some(name) = s.strip_prefix("after:") => {
-                    crate::ast::group::GroupPosition::After(name.to_string())
-                }
-                _ => crate::ast::group::GroupPosition::FirstSymbol,
-            };
-            let spec = crate::ast::group::GroupSpec {
-                module: module.clone(),
-                symbols: sym_names.clone(),
-                preamble: preamble.clone(),
-                position: pos,
-            };
-            let result = crate::ast::group::group_symbols(file_content, &spec, lang_val)?;
-            if result.symbols_moved > 0 {
-                update_file_content(
-                    tx.pending,
-                    tx.deletions,
-                    tx.write_targets,
-                    &abs,
-                    result.content,
-                );
-            }
-            return Ok(result.symbols_moved);
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstMove {
-            path,
-            target,
-            symbols: sym_names,
-            position,
-            target_prepend,
-            lang,
-        } => {
-            let abs_source = tx.cwd.join(path);
-            let abs_target = tx.cwd.join(target);
-            let source_content =
-                read_file_content(tx.pending, tx.existed_before, &abs_source)?.to_string();
-            let target_content = if abs_target.exists() || tx.pending.contains_key(&abs_target) {
-                read_file_content(tx.pending, tx.existed_before, &abs_target)?.to_string()
-            } else {
-                target_prepend
-                    .as_ref()
-                    .map(|p| format!("{p}\n\n"))
-                    .unwrap_or_default()
-            };
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs_source));
-            let pos = crate::ast::move_symbols::parse_position(position.as_deref());
-            let result = crate::ast::move_symbols::move_symbols(
-                &source_content,
-                &target_content,
-                sym_names,
-                pos,
-                lang_val,
-            )?;
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &abs_source,
-                result.source_content,
-            );
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &abs_target,
-                result.target_content,
-            );
-            return Ok(result.symbols_moved);
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstExtractToFile {
-            source,
-            symbol,
-            target,
-            replacement,
-            unwrap,
-            prepend,
-            force,
-            lang,
-        } => {
-            let abs_source = tx.cwd.join(source);
-            let abs_target = tx.cwd.join(target);
-            if !force && (abs_target.exists() || tx.pending.contains_key(&abs_target)) {
-                anyhow::bail!(
-                    "target file '{}' already exists (use force: true to overwrite)",
-                    target
-                );
-            }
-            let source_content = read_file_content(tx.pending, tx.existed_before, &abs_source)?;
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs_source));
-            let do_unwrap = unwrap.unwrap_or(true);
-            let result = crate::ast::extract::extract_to_file(
-                source_content,
-                symbol,
-                replacement.as_deref(),
-                do_unwrap,
-                prepend.as_deref(),
-                lang_val,
-            )?;
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &abs_source,
-                result.source_content,
-            );
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &abs_target,
-                result.target_content,
-            );
-            return Ok(1);
-        }
-
-        #[cfg(feature = "ast")]
-        Operation::AstSplit {
-            source,
-            targets,
-            keep_in_source,
-            source_suffix,
-            source_prefix,
-            require_exhaustive,
-            lang,
-        } => {
-            let abs_source = tx.cwd.join(source);
-            let source_content = read_file_content(tx.pending, tx.existed_before, &abs_source)?;
-            let lang_val = lang
-                .as_deref()
-                .map(crate::ast::Language::from_name_or_ext)
-                .unwrap_or_else(|| crate::ast::Language::from_path(&abs_source));
-            let split_targets: Vec<crate::ast::split::SplitTarget> = targets
-                .iter()
-                .map(|t| crate::ast::split::SplitTarget {
-                    path: t.path.clone(),
-                    symbols: t.symbols.clone(),
-                    prepend: t.prepend.clone(),
-                })
-                .collect();
-            let exhaustive = require_exhaustive.unwrap_or(true);
-            let result = crate::ast::split::split_file(
-                source_content,
-                &split_targets,
-                keep_in_source,
-                source_suffix.as_deref(),
-                source_prefix.as_deref(),
-                exhaustive,
-                lang_val,
-            )?;
-            update_file_content(
-                tx.pending,
-                tx.deletions,
-                tx.write_targets,
-                &abs_source,
-                result.source_content,
-            );
-            for (target_path, target_content) in &result.targets {
-                let abs_target = tx.cwd.join(target_path);
-                update_file_content(
-                    tx.pending,
-                    tx.deletions,
-                    tx.write_targets,
-                    &abs_target,
-                    target_content.clone(),
-                );
-            }
-            return Ok(result.symbols_distributed);
+        Operation::AstRename { .. }
+        | Operation::AstReplace { .. }
+        | Operation::AstInsert { .. }
+        | Operation::AstWrap { .. }
+        | Operation::AstImports { .. }
+        | Operation::AstReorder { .. }
+        | Operation::AstGroup { .. }
+        | Operation::AstMove { .. }
+        | Operation::AstExtractToFile { .. }
+        | Operation::AstSplit { .. } => {
+            return super::ast_op::execute_ast_op(op, tx);
         }
     }
